Add option to decode strings in a reversible way
By default strings are being decoded into unicode in a way that makes it
impossible to encode them back into what's actually in the database.
This causes "data loss" when tables have columns that are supposed to
store raw bytes.

An option "surrogate_encoding=true/false" is added that allows changing
this behavior. The default behavior is kept unchanged for compatibility.